backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.25.0
alembic>=1.13.0
